Add browserstack makefile command
