--- behat/tests/bootstrap/v2/reporting/sections/ContactsFeatureContext.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace DigidepsBehat\v2\Reporting\Sections;
-
-class ContactsFeatureContext extends SectionsBaseFeatureContext
-{
-    use ContactsSectionTrait;
-}
